Move PostgresSQLStoreService code to its own folder
