+++ packages/postgresql-node-connector/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "postgresql-node-connector",
+  "version": "0.0.1",
+  "description": "Implementation of a PostgreSQL storage service for the Counterfactual Node",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "MIT"
+}
