+++ packages/postgresql-node-connector/src/entity/Node.ts
@@ -1,0 +1,10 @@
+import { Column, Entity, PrimaryColumn } from "typeorm";
+
+@Entity()
+export class Node {
+  @PrimaryColumn()
+  key!: string;
+
+  @Column("json")
+  value!: object;
+}
